supply correct date default to Mongoose
The default provided to Mongoose for date values was `Date()`, which produced a single date value corresponding to the start of the program execution. As per the documentation for Mongoose, the function `Date.now` is now provided as the default to the Schema, which ensures that the dates of documents are default initialized to the times at which those documents are constructed.