copy authentication code to routers/auth.js
This is the first step in refactoring all routes. For now, auth.js will
export the loginRouter and logoutRouter, as well as a method that takes in an app object and configures its authentication middleware. Next, these routers must replace the routes for login and logout, and the method will replace the passport initialization. I don't know if this is all valid, so some corrections might be made soon.