Perform several app.js refactors
This commit incorporates the following refactors:
 -logger initialization: moved to logging/logger.js
 -login/logout/signup/signin: moved to various routers in routes/auth.js

It is unclear how to best refactor some of the initializer logic.
Currently, the initialization of Passport is performed by a method
in routes/auth.js that takes in the `app` variable as a parameter.
However, I have yet to encounter this kind of pattern in Express
applications. One might also ask why mounting the routers to the
right endpoints isn't also done in this initializer function of app.js.

It also remains to be seen whether a separate "middleware" directory should
be created.